s2/requirements.txt
fastapi
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
uvicorn[standard]
